chore: rubocop fixes
