+++ lib/integrations/dialogflow/processor_service.rb
@@ -20,7 +20,7 @@ class Integrations::Dialogflow::ProcessorService
   end
 
   def processable_message?(message)
-    #TODO change from reportable and create a dedicated method for this?
+    # TODO: change from reportable and create a dedicated method for this?
     return unless message.reportable?
     return if message.outgoing? && !processable_outgoing_message?(message)
